@@ -98,7 +98,7 @@ brails = [
 ]
 
 [tool.ruff]
-target-version = "py37"
+target-version = "py39"
 line-length = 85
 include = [
     #"brails/types/asset_inventory.py",
